# openapi/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: api@bitly.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://bitly.com/pages/terms-of-service
  title: Bitly API
  version: 4.0.0
  x-logo:
    url: /v4/header-logo.png
servers:
- url: https://api-ssl.bitly.com/v4
security:
- bearerAuth: []
...
paths:
...
  /bitlinks/{bitlink}/countries:
    get:
      description: Returns the country origins of click traffic for the specified
        link.
      operationId: getMetricsForBitlinkByCountries
      parameters:
      - description: A Bitlink made of the domain and hash
        example: bit.ly/12a4b6c
        explode: false
        in: path
        name: bitlink
        required: true
        schema:
          type: string
        style: simple
      - description: A unit of time
        example: month
        explode: true
        in: query
        name: unit
        required: true
        schema:
          default: day
          enum:
          - minute
          - hour
          - day
          - week
          - month
          type: string
        style: form
      - description: An integer representing the time units to query data for. pass
          -1 to return all units of time.
        example: 5
        explode: true
        in: query
        name: units
        required: true
        schema:
          default: -1
          type: integer
        style: form
      - description: The quantity of items to be be returned
        example: 10
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 50
          type: integer
        style: form
      - description: An ISO-8601 timestamp, indicating the most recent time for which
          to pull metrics. Will default to current time.
        example: 2006-01-02T15:04:05-0700
        explode: true
        in: query
        name: unit_reference
        required: false
        schema:
          type: string
        style: form
      responses:
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: BAD_REQUEST
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
          description: NOT_FOUND
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporarilyUnavailable'
          description: TEMPORARILY_UNAVAILABLE
        "200":
          content:
            text/csv:
              example: |
                Country, Clicks
                US, 5
              schema:
                $ref: '#/components/schemas/ClickMetrics'
            application/json:
              schema:
                $ref: '#/components/schemas/ClickMetrics'
          description: SUCCESS
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
          description: INTERNAL_ERROR
      summary: Get Metrics for a Bitlink by Country
      tags:
      - Bitlinks
...
  /user:
    get:
      description: Returns information for the current authenticated user.
      operationId: getUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: SUCCESS
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
          description: NOT_FOUND
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
          description: FORBIDDEN
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
          description: INTERNAL_ERROR
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporarilyUnavailable'
          description: TEMPORARILY_UNAVAILABLE
      summary: Retrieve a User
      tags:
      - User
    patch:
      description: Update fields in the user
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: BAD_REQUEST
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
          description: NOT_FOUND
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
          description: UNPROCESSABLE_ENTITY
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporarilyUnavailable'
          description: TEMPORARILY_UNAVAILABLE
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: SUCCESS
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
          description: INTERNAL_ERROR
      summary: Update a User
      tags:
      - User
...
  /groups/{group_guid}/bitlinks:
    get:
      description: Returns a paginated collection of Bitlinks for a group.
      operationId: getBitlinksByGroup
      parameters:
      - description: A GUID for a Bitly group
        example: Ba1bc23dE4F
        explode: false
        in: path
        name: group_guid
        required: true
        schema:
          type: string
        style: simple
      - description: The quantity of items to be be returned
        example: 10
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 50
          type: integer
        style: form
      - description: Integer specifying the numbered result at which to start
        example: 1
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - description: Custom keyword to filter on history entries
        example: chauncey
        explode: true
        in: query
        name: keyword
        required: false
        schema:
          type: string
        style: form
      - description: the value that you would like to search
        example: api
        explode: true
        in: query
        name: query
        required: false
        schema:
          type: string
        style: form
      - description: Timestamp as an integer unix epoch
        example: 1501027200
        explode: true
        in: query
        name: created_before
        required: false
        schema:
          type: integer
        style: form
      - description: Timestamp as an integer unix epoch
        example: 1501027200
        explode: true
        in: query
        name: created_after
        required: false
        schema:
          type: integer
        style: form
      - description: Timestamp as an integer unix epoch
        example: 1501027200
        explode: true
        in: query
        name: modified_after
        required: false
        schema:
          type: integer
        style: form
      - description: Whether or not to include archived bitlinks
        example: both
        explode: true
        in: query
        name: archived
        required: false
        schema:
          default: "off"
          enum:
          - "on"
          - "off"
          - both
          type: string
        style: form
      - description: Filter to only Bitlinks that contain deeplinks
        example: both
        explode: true
        in: query
        name: deeplinks
        required: false
        schema:
          default: both
          enum:
          - "on"
          - "off"
          - both
          type: string
        style: form
      - description: Filter to only Bitlinks that contain deeplinks configured with
          a custom domain
        example: both
        explode: true
        in: query
        name: domain_deeplinks
        required: false
        schema:
          default: both
          enum:
          - "on"
          - "off"
          - both
          type: string
        style: form
      - description: Filter to return only links for the given campaign GUID, can
          be provided
        example: Ca1bcd2EFGh
        explode: true
        in: query
        name: campaign_guid
        required: false
        schema:
          type: string
        style: form
      - description: Filter to return only links for the given channel GUID, can be
          provided, overrides all other parameters
        example: Ha1bc2DefGh
        explode: true
        in: query
        name: channel_guid
        required: false
        schema:
          type: string
        style: form
      - example: both
        explode: true
        in: query
        name: custom_bitlink
        required: false
        schema:
          default: both
          enum:
          - "on"
          - "off"
          - both
          type: string
        style: form
      - description: filter by given tags
        example:
        - bitly
        - api
        explode: true
        in: query
        name: tags
        required: false
        schema:
          items:
            type: string
          type: array
        style: form
      - description: filter by launchpad id
        example:
        - M1234567890
        explode: true
        in: query
        name: launchpad_ids
        required: false
        schema:
          items:
            type: string
          type: array
        style: form
      - description: Filter by the login of the authenticated user that created the
          Bitlink
        example:
        - chauncey
        explode: true
        in: query
        name: encoding_login
        required: false
        schema:
          items:
            type: string
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bitlinks'
          description: SUCCESS
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
          description: NOT_FOUND
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
          description: FORBIDDEN
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
          description: INTERNAL_ERROR
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporarilyUnavailable'
          description: TEMPORARILY_UNAVAILABLE
      summary: Retrieve Bitlinks by Group
      tags:
      - Bitlinks
...
components:
...
  schemas:
...
    ClickMetric:
      properties:
        clicks:
          type: integer
        value:
          type: string
      type: object
...
    HasReferences:
      properties:
        references:
          additionalProperties:
            type: string
          type: object
      type: object
...
    BaseMetrics:
      properties:
        units:
          type: integer
        facet:
          enum:
          - countries
          - cities
          - devices
          - referrers
          - referrers_by_domain
          - referring_domains
          - referring_networks
          - shorten_counts
          type: string
        unit_reference:
          type: string
        unit:
          type: string
      type: object
...
    Email:
      example:
        is_primary: true
        is_verified: true
        email: email
      properties:
        is_primary:
          type: boolean
        is_verified:
          type: boolean
        email:
          type: string
      type: object
...
    Pagination:
      example:
        next: next
        total: 0
        size: 6
        prev: prev
        page: 1
      properties:
        total:
          type: integer
        size:
          type: integer
        prev:
          type: string
        page:
          type: integer
        next:
          type: string
      type: object
...
    UserUpdate:
      example:
        default_group_guid: Ba1bc23dE4F
        name: Chauncey McPufferson
      properties:
        default_group_guid:
          type: string
        name:
          type: string
      type: object
...
    BitlinkUpdate:
      properties:
        archived:
          type: boolean
        tags:
          items:
            type: string
          type: array
        created_at:
          type: string
        title:
          type: string
        deeplinks:
          items:
            $ref: '#/components/schemas/DeeplinkRule'
          type: array
        created_by:
          type: string
        long_url:
          type: string
        client_id:
          type: string
        custom_bitlinks:
          items:
            type: string
          type: array
        link:
          type: string
        launchpad_ids:
          items:
            type: string
          type: array
        id:
          type: string
      type: object
...
    Bitlinks:
      example:
        pagination:
          next: next
          total: 0
          size: 6
          prev: prev
          page: 1
        links:
        - null
        - null
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        links:
          items:
            $ref: '#/components/schemas/BitlinkBody'
          type: array
      type: object
...
    User:
      example:
        emails:
        - is_primary: true
          is_verified: true
          email: email
        - is_primary: true
          is_verified: true
          email: email
        is_2fa_enabled: true
        is_active: true
        created: created
        default_group_guid: default_group_guid
        name: name
        modified: modified
        login: login
        is_sso_user: true
      properties:
        default_group_guid:
          type: string
        name:
          type: string
        created:
          type: string
        is_active:
          type: boolean
        modified:
          type: string
        is_sso_user:
          type: boolean
        is_2fa_enabled:
          type: boolean
        login:
          type: string
        emails:
          items:
            $ref: '#/components/schemas/Email'
          type: array
      required:
      - created
      - emails
      - is_2fa_enabled
      - is_active
      - is_sso_user
      - login
      - modified
      - name
      type: object
...
    NotFound:
      allOf:
      - $ref: '#/components/schemas/Error'
      description: NOT_FOUND
...
    TemporarilyUnavailable:
      allOf:
      - $ref: '#/components/schemas/Error'
      description: TEMPORARILY_UNAVAILABLE
...
    Forbidden:
      allOf:
      - $ref: '#/components/schemas/Error'
      description: FORBIDDEN
...
    InternalError:
      allOf:
      - $ref: '#/components/schemas/Error'
      description: INTERNAL_ERROR
...
    Error:
      properties:
        message:
          type: string
        errors:
          items:
            $ref: '#/components/schemas/FieldError'
          type: array
        resource:
          type: string
        description:
          type: string
      type: object
    ClickMetrics:
      allOf:
      - $ref: '#/components/schemas/BaseMetrics'
      - $ref: '#/components/schemas/ClickMetrics_allOf'
      type: object
    BadRequest:
      allOf:
      - $ref: '#/components/schemas/Error'
      description: BAD_REQUEST
...
    BitlinkBody:
      allOf:
      - $ref: '#/components/schemas/HasReferences'
      - $ref: '#/components/schemas/BitlinkUpdate'
      type: object
...
    FieldError:
      properties:
        field:
          type: string
        message:
          type: string
        error_code:
          type: string
      type: object
...
    UnprocessableEntity:
      allOf:
      - $ref: '#/components/schemas/Error'
      description: UNPROCESSABLE_ENTITY
...
    DeeplinkRule:
      properties:
        bitlink:
          type: string
        install_url:
          type: string
        created:
          description: ISO timestamp
          type: string
        app_uri_path:
          type: string
        modified:
          description: ISO timestamp
          type: string
        install_type:
          type: string
        app_guid:
          type: string
        guid:
          type: string
        os:
          type: string
        brand_guid:
          type: string
      type: object
...
    ClickMetrics_allOf:
      properties:
        metrics:
          items:
            $ref: '#/components/schemas/ClickMetric'
          type: array
...
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http